tie in planner changes with the new pkg
