+++ pkg/blockbuilder/scheduler/kafkautil_test.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package blockscheduler
+package scheduler
 
 import (
 	"context"
