+++ pkg/blockbuilder/scheduler/scheduler_test.go
@@ -5,20 +5,23 @@ import (
 	"testing"
 	"time"
 
+	"github.com/go-kit/log"
+	"github.com/prometheus/client_golang/prometheus"
+
 	"github.com/grafana/loki/v3/pkg/blockbuilder/builder"
 	"github.com/grafana/loki/v3/pkg/blockbuilder/types"
 )
 
 type testEnv struct {
 	queue     *JobQueue
-	scheduler *QueueScheduler
+	scheduler *BlockScheduler
 	transport *builder.MemoryTransport
 	builder   *builder.Worker
 }
 
 func newTestEnv(builderID string) *testEnv {
 	queue := NewJobQueue()
-	scheduler := NewScheduler(queue)
+	scheduler := NewScheduler(Config{}, queue, nil, log.NewNopLogger(), prometheus.NewRegistry())
 	transport := builder.NewMemoryTransport(scheduler)
 	builder := builder.NewWorker(builderID, builder.NewMemoryTransport(scheduler))
